requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
pynput>=1.7.6
pymysql>=1.1.0
httpx>=0.27.0
keyring>=25.0.0

# v3.4 来源 App 捕获（按平台可选）
pywin32 >= 306; sys_platform == "win32"
psutil >= 5.9; sys_platform == "win32" or sys_platform == "linux"
pyobjc-framework-Cocoa >= 10.0; sys_platform == "darwin"
python-xlib >= 0.33; sys_platform == "linux"
jeepney >= 0.8; sys_platform == "linux"
